requirements.txt
gitpython==3.1.31
flask==2.3.2
flask-wtf==1.1.1
pyyaml==6.0
python-dateutil==2.8.2
watchdog==3.0.0 